pyproject.toml
[project]
name = "sec-gazette"
version = "0.1.0"
description = "SEC News Digest extraction pipeline for academic research on audit quality"
package-mode = false
authors = [
    {name = "CasparDP",email = "58906924+CasparDP@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "duckdb (>=1.4.3,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "docling>=2.0.0",
    "ollama>=0.4.0",
    "lxml>=5.3.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
